add/sub -> rtype gadget to build circuit
